Modified code to work with pythonanywhere
